+++ better-payment/includes/Admin/views/partials/layout-repeater-vars.php
@@ -51,7 +51,7 @@ if ($is_payment_amount_field) {
     $required_placeholder = ' *';
     
     $render_attribute_default_dynamic = ! empty( $item["better_payment_field_name_default_dynamic_enable"] ) && 'yes' ===  $item["better_payment_field_name_default_dynamic_enable"] ? 1 : 0;
-    $render_attribute_default = $render_attribute_default_dynamic && ! empty( $_GET['payment_amount'] ) ? intval($_GET['payment_amount']) : $render_attribute_default;
+    $render_attribute_default = $render_attribute_default_dynamic && ! empty( $_GET['payment_amount'] ) ? floatval($_GET['payment_amount']) : $render_attribute_default;
 }
 
 if ($is_primary_email_field) {
